Move Employee and utilities to their files
